models.BlogPost: shift initing of deps into get_deps
This fixes calls to BlogPost.get_deps() from
PostRegenerator.regenerate().

We could have added the check-and-init-if-needed code, as we do in
post.publish() and post.remove(), but we shift this into post.get_deps()
itself. This way, callers of get_deps() don't have to remember to check
and init post.deps.